requirements.txt
discord.py
python-dotenv
requests
google-cloud-storage
selenium
webdriver_manager